feat(frameParser): adds methods creating settings frame
- refactors the FrameLenght field of frames
- adds function to calculate expected lenght after varint encoding
- adds enum value for settings in settings frame
- function for creating headers and settings frame, preparing the data
  to before being encoded

+++ pkg/frameParser/testLab.go
@@ -80,7 +80,7 @@ func testDataFrame() {
 func testSettingsFrame() {
 	// create SettingsFrame with key values pairs
 	originalFrame := &SettingsFrame{
-		Settings: map[uint16]uint64{
+		Settings: map[uint64]uint64{
 			0x1: 0x100,
 			0x2: 0x200,
 		},
